clean up pr
