+++ apps/src/aiDifferentiation/personalization/PersonalizationResults.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import {TeachingStyle} from './personalization_types';
 import PersonalizationResultsColumnAiHelp from './PersonalizationResultsColumnAiSupport';
 import PersonalizationResultsColumnArrows from './PersonalizationResultsColumnArrows';
 import PersonalizationResultsColumnSuperpowers from './PersonalizationResultsColumnSuperpowers';
@@ -7,15 +8,6 @@ import PersonalizationResultsHeader from './PersonalizationResultsHeader';
 import PersonalizationResultsInfoBox from './PersonalizationResultsInfoBox';
 
 import style from './personalization-information.module.scss';
-
-// should this be somewhere else?
-export interface TeachingStyle {
-  name: string;
-  emoji: string;
-  tagline: string;
-  teachingSuperpowers: string[];
-  howAiHelps?: string[];
-}
 interface PersonalizationResultsProps {
   teachingStyle: TeachingStyle;
 }
