+++ apps/src/aiDifferentiation/personalization/personalization_types.ts
@@ -1,0 +1,8 @@
+// should this be somewhere else?
+export interface TeachingStyle {
+  name: string;
+  emoji: string;
+  tagline: string;
+  teachingSuperpowers: string[];
+  howAiHelps?: string[];
+}
